feat(onchain_queries): Plumbing
