+++ pragma-node/src/handlers/entries/create_entry.rs
@@ -2,15 +2,39 @@ use axum::extract::State;
 use axum::Json;
 use chrono::{DateTime, Utc};
 use pragma_entities::{EntryError, NewEntry, PublisherError};
+use serde::{Deserialize, Serialize};
 use starknet::core::types::FieldElement;
+use utoipa::ToSchema;
 
 use crate::config::config;
-use crate::handlers::entries::{CreateEntryRequest, CreateEntryResponse};
 use crate::infra::kafka;
 use crate::infra::repositories::publisher_repository;
 use crate::types::entries::Entry;
 use crate::utils::{assert_request_signature_is_valid, JsonExtractor};
 use crate::AppState;
+
+#[derive(Debug, Serialize, Deserialize, ToSchema)]
+pub struct CreateEntryRequest {
+    pub signature: Vec<FieldElement>,
+    pub entries: Vec<Entry>,
+}
+
+impl AsRef<[FieldElement]> for CreateEntryRequest {
+    fn as_ref(&self) -> &[FieldElement] {
+        &self.signature
+    }
+}
+
+impl AsRef<[Entry]> for CreateEntryRequest {
+    fn as_ref(&self) -> &[Entry] {
+        &self.entries
+    }
+}
+
+#[derive(Debug, Serialize, Deserialize, ToSchema)]
+pub struct CreateEntryResponse {
+    number_entries_created: usize,
+}
 
 #[utoipa::path(
     post,
