+++ pragma-node/src/handlers/entries/get_onchain/publishers.rs
@@ -1,14 +1,47 @@
 use axum::extract::{Query, State};
 use axum::Json;
 
+use pragma_common::types::{DataType, Network};
 use pragma_entities::EntryError;
+use serde::{Deserialize, Serialize};
+use utoipa::{IntoParams, ToSchema};
 
-use crate::handlers::entries::{GetOnchainPublishersParams, GetOnchainPublishersResponse};
 use crate::infra::repositories::entry_repository::get_all_currencies_decimals;
 use crate::infra::repositories::onchain_repository::{
     get_publishers, get_publishers_with_components,
 };
 use crate::AppState;
+
+#[derive(Debug, Default, Deserialize, IntoParams, ToSchema)]
+pub struct GetOnchainPublishersParams {
+    pub network: Network,
+    pub data_type: DataType,
+}
+
+#[derive(Debug, Serialize, Deserialize, ToSchema)]
+pub struct PublisherEntry {
+    pub pair_id: String,
+    pub last_updated_timestamp: u64,
+    pub price: String,
+    pub source: String,
+    pub decimals: u32,
+    pub daily_updates: u32,
+}
+
+#[derive(Debug, Serialize, Deserialize, ToSchema)]
+pub struct Publisher {
+    pub publisher: String,
+    pub website_url: String,
+    pub last_updated_timestamp: u64,
+    pub r#type: u32,
+    pub nb_feeds: u32,
+    pub daily_updates: u32,
+    pub total_updates: u32,
+    pub components: Vec<PublisherEntry>,
+}
+
+#[derive(Debug, Default, Serialize, Deserialize, ToSchema)]
+pub struct GetOnchainPublishersResponse(pub Vec<Publisher>);
 
 #[utoipa::path(
     get,
