+++ pragma-node/src/servers/app.rs
@@ -45,25 +45,24 @@ pub async fn run_app_server(config: &Config, state: AppState) {
             schemas(pragma_entities::dto::Publisher, pragma_entities::PublisherError),
             schemas(pragma_entities::error::InfraError),
             schemas(
-                handlers::entries::CreateEntryRequest,
-                handlers::entries::CreateEntryResponse,
-                handlers::entries::CreateFutureEntryRequest,
-                handlers::entries::CreateFutureEntryResponse,
+                handlers::entries::create_entry::CreateEntryRequest,
+                handlers::entries::create_entry::CreateEntryResponse,
+                handlers::entries::create_future_entry::CreateFutureEntryRequest,
+                handlers::entries::create_future_entry::CreateFutureEntryResponse,
                 handlers::entries::GetEntryParams,
-                handlers::entries::GetEntryResponse,
-                handlers::entries::SubscribeToEntryResponse,
-                handlers::entries::GetVolatilityResponse,
-                handlers::entries::GetOHLCResponse,
-                handlers::entries::GetOnchainParams,
-                handlers::entries::GetOnchainResponse,
-                handlers::entries::GetOnchainCheckpointsParams,
-                handlers::entries::GetOnchainCheckpointsResponse,
-                handlers::entries::GetOnchainPublishersParams,
-                handlers::entries::GetOnchainPublishersResponse,
-                handlers::entries::GetOnchainOHLCResponse,
-                handlers::entries::GetOnchainHistoryParams,
-                handlers::entries::GetOnchainHistoryEntry,
-                handlers::entries::GetOnchainHistoryResponse,
+                handlers::entries::get_entry::GetEntryResponse,
+                handlers::entries::subscribe_to_entry::SubscribeToEntryResponse,
+                handlers::entries::get_volatility::GetVolatilityResponse,
+                handlers::entries::get_ohlc::GetOHLCResponse,
+                handlers::entries::get_onchain::GetOnchainParams,
+                handlers::entries::get_onchain::GetOnchainResponse,
+                handlers::entries::get_onchain::checkpoints::GetOnchainCheckpointsParams,
+                handlers::entries::get_onchain::checkpoints::GetOnchainCheckpointsResponse,
+                handlers::entries::get_onchain::publishers::GetOnchainPublishersParams,
+                handlers::entries::get_onchain::publishers::GetOnchainPublishersResponse,
+                handlers::entries::get_onchain::ohlc::GetOnchainOHLCResponse,
+                handlers::entries::get_onchain::history::GetOnchainHistoryParams,
+                handlers::entries::get_onchain::history::GetOnchainHistoryResponse,
 
             ),
             schemas(
@@ -71,10 +70,11 @@ pub async fn run_app_server(config: &Config, state: AppState) {
                 types::entries::Entry,
                 types::entries::PerpEntry,
                 types::entries::FutureEntry,
-                handlers::entries::OnchainEntry,
-                handlers::entries::Checkpoint,
-                handlers::entries::Publisher,
-                handlers::entries::PublisherEntry,
+                handlers::entries::get_onchain::OnchainEntry,
+                handlers::entries::get_onchain::checkpoints::Checkpoint,
+                handlers::entries::get_onchain::publishers::Publisher,
+                handlers::entries::get_onchain::publishers::PublisherEntry,
+                handlers::entries::get_onchain::history::GetOnchainHistoryEntry,
             ),
             schemas(
                 pragma_common::types::AggregationMode,
